Bug 1808801 - Update Test262 January 5, 2023 r=arai
Differential Revision: https://phabricator.services.mozilla.com/D166363

UltraBlame original commit: 123149ff06449baf9c697c9ed97a2d82ea91170e

+++ js/src/tests/test262/intl402/DurationFormat/constructor-options-style-valid.js
@@ -14,7 +14,7 @@
 
 
 const validOptions = [
-  [undefined, "long"],
+  [undefined, "short"],
   ["long", "long"],
   ["short", "short"],
   ["narrow", "narrow"],
